Separated navigators & switched to named functions
The stack navigator and tab navigator have been separated and can now be found within the app/navigation folder. There are functions/state in TabNavigator.js that need to be moved to the redux actions folder.

As a minor fix, all components and screens have adopted the named function convention to be consistent with the rest of the application.
