--- app/assets/config/colors.js
@@ -1,15 +1,0 @@
-export default {
-  primary: '#fff', //white
-  buttons: '#000', //black
-  background: 'teal',
-  buttonText: '#fff',//white
-  textInput: '#fff',
-  medium: '#6e6969',//medium gray
-  dark: '#0c0c0c',//dark gray
-  white: '#fff',
-  blue: '#0000FF',
-  lightgray: '#00000040',
-  tabs: '#99e0c8'//light green used on the tabs
-
-  //we can edit these colors once the wireframes are done.
-};
