+++ app/assets/theme/colors.js
@@ -1,0 +1,13 @@
+export default {
+  primary: '#fff', // White
+  buttons: '#000', // Black
+  background: 'teal',
+  buttonText: '#fff', // White
+  textInput: '#fff',
+  medium: '#6e6969', // Medium gray
+  dark: '#0c0c0c', // Dark gray
+  white: '#fff',
+  blue: '#0000FF',
+  lightgray: '#00000040',
+  tabs: '#99e0c8', // Light green
+};
